refactor: added comparisons.rs
